+++ cartridge/confapplier.lua
@@ -71,20 +71,19 @@ local state_transitions = {
     -- Remote control is running.
     -- Clusterwide config is loaded.
     -- Remote control initiated `boot_instance()`
-    ['BootstrappingBox'] = {'BoxConfigured', 'BootError'},
+    ['BootstrappingBox'] = {'ConnectingFullmesh', 'BootError'},
 
     -- Remote control is running.
     -- Clusterwide config is loaded.
     -- Function `confapplier.init()` initiated `boot_instance()`
-    ['RecoveringSnapshot'] = {'BoxConfigured', 'BootError'},
+    ['RecoveringSnapshot'] = {'ConnectingFullmesh', 'BootError'},
 
     -- Remote control is stopped.
     -- Recovering snapshot finished.
     -- Box is listening binary port.
-    ['BoxConfigured'] = {'ConnectingFullmesh'},
+    ['ConnectingFullmesh'] = {'ConfiguringRoles'},
 
 -- normal operation
-    ['ConnectingFullmesh'] = {'ConfiguringRoles', 'OperationError'},
     ['ConfiguringRoles'] = {'RolesConfigured', 'OperationError'},
     ['RolesConfigured'] = {'ConfiguringRoles'},
 
@@ -193,10 +192,13 @@ local function validate_config(clusterwide_config, _)
 end
 
 
+local configure_fullmesh
+local configure_roles
+
 --- Apply the role configuration.
 -- @function apply_config
 -- @local
--- @tparam table conf
+-- @tparam table clusterwide_config
 -- @treturn[1] boolean true
 -- @treturn[2] nil
 -- @treturn[2] table Error description
@@ -204,16 +206,11 @@ local function apply_config(clusterwide_config)
     checks('ClusterwideConfig')
     assert(clusterwide_config.locked)
     assert(
-        vars.state == 'BoxConfigured'
-        or vars.state == 'RolesConfigured',
+        vars.state == 'RolesConfigured',
         'Unexpected state ' .. vars.state
     )
 
     vars.clusterwide_config = clusterwide_config
-
-    if vars.state == 'BoxConfigured' then
-        set_state('ConnectingFullmesh')
-    end
 
     local _, err = BoxError:pcall(box.cfg, {
         replication = topology.get_fullmesh_replication(
@@ -223,17 +220,57 @@ local function apply_config(clusterwide_config)
     if err ~= nil then
         set_state('OperationError', err)
         return nil, err
-    elseif box.info.status == 'orphan' then
+    end
+
+    local _, err = configure_fullmesh()
+    if err ~= nil then
+        set_state('OperationError', err)
+        return nil, err
+    end
+
+    return configure_roles(clusterwide_config)
+end
+
+--- Confiure fullmesh replication
+-- @configure_fullmesh
+-- @local
+-- @treturn[1] boolean true
+-- @treturn[2] nil
+-- @treturn[2] table Error description
+configure_fullmesh = function()
+    assert(
+        vars.state == 'ConnectingFullmesh'
+        or vars.state == 'RolesConfigured',
+        'Unexpected state ' .. vars.state
+    )
+
+    if box.info.status == 'orphan' then
         local err = BoxError:new(
             'Replication setup failed, instance in orphan mode'
         )
-        set_state('OperationError', err)
         return nil, err
-    else
-        box.cfg({replication_connect_quorum = 0})
     end
 
+    box.cfg({replication_connect_quorum = 0})
     set_state('ConfiguringRoles')
+    return true
+end
+
+
+--- Configuring roles
+-- @configure_roles
+-- @local
+-- @tparam table clusterwide_config
+-- @treturn[1] boolean true
+-- @treturn[2] nil
+-- @treturn[2] table Error description
+configure_roles = function(clusterwide_config)
+    checks('ClusterwideConfig')
+    assert(clusterwide_config.locked)
+    assert(
+        vars.state =='ConfiguringRoles',
+        'Unexpected state ' .. vars.state
+    )
     failover.cfg(clusterwide_config)
 
     local ok, err = ddl_manager.apply_config(
@@ -254,7 +291,6 @@ local function apply_config(clusterwide_config)
     set_state('RolesConfigured')
     return true
 end
-
 
 local function boot_instance(clusterwide_config)
     checks('ClusterwideConfig')
@@ -465,8 +501,32 @@ local function boot_instance(clusterwide_config)
         return nil, err
     end
 
-    set_state('BoxConfigured')
-    return apply_config(clusterwide_config)
+    local _, err = BoxError:pcall(box.cfg, {
+        replication = topology.get_fullmesh_replication(
+            clusterwide_config:get_readonly('topology'), vars.replicaset_uuid
+        ),
+    })
+    if err ~= nil then
+        set_state('BootError', err)
+        return nil, err
+    end
+
+    vars.clusterwide_config = clusterwide_config
+    set_state('ConnectingFullmesh')
+
+    local res, err = configure_fullmesh()
+    -- Waiting untill fullmesh replocication become alive
+    if err ~= nil then
+        fiber.new(function()
+            while err ~= nil do
+                res, err = configure_fullmesh()
+                fiber.sleep(1)
+            end
+            return configure_roles(clusterwide_config)
+        end)
+        return nil, err
+    end
+    return configure_roles(clusterwide_config)
 end
 
 local function init(opts)