+++ test/integration/state_machine_test.lua
@@ -377,7 +377,7 @@ end
 function g.test_orphan_connect_timeout()
     -- If the master can't connect to the slave in
     -- <TARANTOOL_REPLICATION_CONNECT_TIMEOUT> seconds
-    -- it becomes an orphan and transits to OperationError state.
+    -- it stays in ConnectingFullmesh state until reconnects
 
     g.master:stop()
     g.slave:stop()
@@ -398,7 +398,7 @@ function g.test_orphan_connect_timeout()
             message = "",
         }}
     )
-    wish_state(g.master, 'OperationError')
+    wish_state(g.master, 'ConnectingFullmesh')
 
     t.helpers.retrying({}, function()
         t.assert_equals(list_issues(g.master), {{
@@ -413,6 +413,7 @@ function g.test_orphan_connect_timeout()
     log.info('--------------------------------------------------------')
     g.slave:start()
     wish_state(g.slave, 'RolesConfigured')
+    wish_state(g.master, 'RolesConfigured')
 
     t.assert_equals(
         get_upstream_info(g.slave),
@@ -423,9 +424,11 @@ function g.test_orphan_connect_timeout()
         }
     )
 
-    t.assert_equals(rpc_get_candidate(g.slave), g.slave.advertise_uri)
-    t.assert_equals(get_leader(g.slave), g.slave.instance_uuid)
-    t.assert_equals(is_master(g.slave), true)
+    t.assert_equals(rpc_get_candidate(g.slave), g.master.advertise_uri)
+    t.assert_equals(get_leader(g.slave), g.master.instance_uuid)
+    t.assert_equals(is_master(g.slave), false)
+    t.assert_equals(is_master(g.master), true)
+
     t.helpers.retrying({}, function()
         t.assert_equals(list_issues(g.slave), {})
     end)
@@ -441,7 +444,7 @@ function g.test_orphan_sync_timeout()
     g.master.env['TARANTOOL_REPLICATION_SYNC_LAG'] = 1e-308
     g.master.env['TARANTOOL_REPLICATION_SYNC_TIMEOUT'] = 0.1
     g.master:start()
-    wish_state(g.master, 'OperationError')
+    wish_state(g.master, 'ConnectingFullmesh')
 
     t.assert_equals(
         -- master <- slave replication is established instantly